+++ src/App.js
@@ -1,8 +1,23 @@
-import { Outlet } from 'react-router-dom';
+import {
+  Outlet,
+  createBrowserRouter,
+  RouterProvider,
+  useParams,
+} from 'react-router-dom';
+import ErrorPage from './routes/error-page';
 import Navbar from './components/Navbar/Navbar';
+import './index.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import Categories from './components/Categories/Categories';
+import MusicList from './components/MusicList/MusicList';
 
 function App() {
-  return (
+  const GetMusiList = () => {
+    const { id } = useParams();
+    return <MusicList category={id} />;
+  };
+
+  const View = () => (
     <div className="App">
       <Navbar />
       <div className="container body">
@@ -12,6 +27,29 @@ function App() {
         <p>Copyright &copy; 2023 Chudisoft. All rights reserved.</p>
       </footer>
     </div>
+  );
+
+  const router = createBrowserRouter([
+    {
+      path: '/',
+      element: <View />,
+      errorElement: <ErrorPage />,
+      children: [
+        {
+          path: '/',
+          element: <Categories />,
+        },
+        {
+          path: '/Music/:id',
+          element: <GetMusiList />,
+          // element: <MusicList category="pop-music" />,
+        },
+      ],
+    },
+  ]);
+
+  return (
+    <RouterProvider router={router} />
   );
 }
