+++ src/index.js
@@ -1,47 +1,12 @@
 import * as ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider, useParams } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import ErrorPage from './routes/error-page';
 import store from './redux/store';
-import './index.css';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import Categories from './components/Categories/Categories';
-import MusicList from './components/MusicList/MusicList';
-
-const GetMusiList = () => {
-  const { id } = useParams();
-  if (id === '') return <MusicList category={id} />;
-  return <Categories />;
-};
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    errorElement: <ErrorPage />,
-    children: [
-      {
-        path: '/',
-        element: <Categories />,
-      },
-      {
-        path: '/music/:id',
-        element: <GetMusiList />,
-        // element: <MusicList category="pop-music" />,
-      },
-      // {
-      //   path: '/profile',
-      //   element: <Profile />,
-      // },
-    ],
-  },
-]);
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <Provider store={store}>
-    <RouterProvider router={router} />
+    <App />
   </Provider>,
 );
